+++ src/librand/lib.rs
@@ -29,6 +29,7 @@
 #![unstable(feature = "rand",
             reason = "use `rand` from crates.io")]
 #![feature(core_float)]
+#![feature(core_intrinsics)]
 #![feature(core_slice_ext)]
 #![feature(no_std)]
 #![feature(num_bits_bytes)]
@@ -42,6 +43,8 @@
 #[cfg(test)] #[macro_use] extern crate std;
 #[cfg(test)] #[macro_use] extern crate log;
 
+use core::f64;
+use core::intrinsics;
 use core::marker::PhantomData;
 
 pub use isaac::{IsaacRng, Isaac64Rng};
@@ -58,6 +61,43 @@ pub mod isaac;
 pub mod chacha;
 pub mod reseeding;
 mod rand_impls;
+
+// Temporary trait to implement a few floating-point routines
+// needed by librand; this is necessary because librand doesn't
+// depend on libstd.  This will go away when librand is integrated
+// into libstd.
+trait FloatMath : Sized {
+    fn exp(self) -> Self;
+    fn ln(self) -> Self;
+    fn sqrt(self) -> Self;
+    fn powf(self, n: Self) -> Self;
+}
+
+impl FloatMath for f64 {
+    #[inline]
+    fn exp(self) -> f64 {
+        unsafe { intrinsics::expf64(self) }
+    }
+
+    #[inline]
+    fn ln(self) -> f64 {
+        unsafe { intrinsics::logf64(self) }
+    }
+
+    #[inline]
+    fn powf(self, n: f64) -> f64 {
+        unsafe { intrinsics::powf64(self, n) }
+    }
+
+    #[inline]
+    fn sqrt(self) -> f64 {
+        if self < 0.0 {
+            f64::NAN
+        } else {
+            unsafe { intrinsics::sqrtf64(self) }
+        }
+    }
+}
 
 /// A type that can be randomly generated using an `Rng`.
 #[doc(hidden)]
